Refactored projects page.
