+++ client/data/caseStudies.js
@@ -1,69 +1,67 @@
 export const preview = [
   {
-    projectName: 'nambafood',
-    logoImg: 'nambafood-icon',
+    name: 'nambafood',
+    logo: 'nambafood-icon',
     logoWidth: 104,
-    maddevsLogo: 'md-logo-black',
+    mdLogo: 'md-logo-black',
     link: '/case-studies/namba-food/',
-    targetBlank: false,
-    projectBackground: 'nambafoodBackground',
-    fileExtension: 'png',
-    projectColor: '#f8f7f6',
-    projectTitle: 'Contactless Delivery Service',
-    projectDescription:
+    isBlank: false,
+    background: 'nambafoodBackground',
+    extension: 'png',
+    color: '#f8f7f6',
+    title: 'Contactless Delivery Service',
+    description:
       'Mad Devs created the Namba Food delivery service from scratch. The solution orchestrates feature-rich apps for couriers, end-users, and business owners.',
-
     contribution: 'Backend, Infrastructure, Frontend',
     alt: 'Namba Food Contactless Delivery Service`s Logo.',
   },
   {
-    projectName: 'sjmc',
-    logoImg: 'sjmc-icon',
+    name: 'sjmc',
+    logo: 'sjmc-icon',
     logoWidth: 91.5,
-    maddevsLogo: 'md-logo-white',
+    mdLogo: 'md-logo-white',
     link: '/case-studies/sir-john-monash-centre/',
-    targetBlank: false,
-    projectBackground: 'sjmcBackground',
-    fileExtension: 'jpg',
-    projectColor: '#221F20',
-    projectTitle: 'Australian National Interpretive Centre',
-    projectDescription:
+    isBlank: false,
+    background: 'sjmcBackground',
+    extension: 'jpg',
+    color: '#221F20',
+    title: 'Australian National Interpretive Centre',
+    description:
       'Mad Devs engineers helped Sir John Monash Centre to enhance and maintain the BYOD multimedia technology.',
 
     contribution: 'Backend, Infrastructure, Frontend',
     alt: 'General John Monash The Australian War Memorial`s Logo.',
   },
   {
-    projectName: 'guardrails',
-    logoImg: 'guardrails-icon',
+    name: 'guardrails',
+    logo: 'guardrails-icon',
     logoWidth: 174,
-    maddevsLogo: 'md-logo-white',
+    mdLogo: 'md-logo-white',
     link:
       'https://blog.maddevs.io/guardrails-helps-800-development-teams-to-establish-devsecops-with-automated-security-reviews-806c3c1d516b?source=friends_link&sk=15f4cdb98f9410b27b9d76dfd94ad1d0',
-
-    targetBlank: true,
-    projectBackground: 'guardrailsBackground',
-    fileExtension: 'png',
-    projectColor: '#0e1b27',
-    projectTitle: 'Cloud cybersecurity service',
-    projectDescription:
+    isBlank: true,
+    background: 'guardrailsBackground',
+    extension: 'png',
+    color: '#0e1b27',
+    title: 'Cloud cybersecurity service',
+    description:
       "Mad Devs was involved with Guardrails' security check service as a development contractor with exceptional knowledge of GitHub and GitLab processes.",
 
     contribution: 'Backend, Infrastructure, Frontend',
     alt: 'Guardrails Cloud CyberSecurity Services Logo.',
   },
   {
-    projectName: 'godee',
-    logoImg: 'godee-icon',
+    name: 'godee',
+    logo: 'godee-icon',
     logoWidth: 121,
-    maddevsLogo: 'md-logo-black',
+    mdLogo: 'md-logo-black',
     link: '/case-studies/godee/',
-    targetBlank: false,
-    projectBackground: 'godeeBackground',
-    fileExtension: 'png',
-    projectColor: '#ff6A01',
-    projectTitle: 'Convenient shuttle bus service',
-    projectDescription:
+    isBlank: false,
+    background: 'godeeBackground',
+    extension: 'png',
+    color: '#ff6A01',
+    title: 'Convenient shuttle bus service',
+    description:
       'Mad Devs helped GoDee with developing feature-rich software to re-invent public mobility by building new smart ways of a daily commute.',
 
     contribution: 'Backend, Infrastructure, Mobile Apps',
