+++ client/data/openSource.js
@@ -1,0 +1,34 @@
+export const projects = [
+  {
+    name: 'mad-location',
+    link: 'https://github.com/maddevsio/mad-location-manager',
+    description: 'А library for GPS and accelerometer data "fusion" with a Kalman filter.',
+  },
+  {
+    name: 'comedian',
+    link: 'https://github.com/maddevsio/comedian',
+    description:
+      'A <span>team-management</span> system leveraging Slack and Telegram bot functionalities to enable remote standups and track daily performance.',
+  },
+  {
+    name: 'ariadna',
+    link: 'https://github.com/maddevsio/ariadna',
+    description:
+      'An <span>open-source</span> geocoder, built on top of ElasticSearch, for fast geocoding and better search for CIS countries.',
+  },
+  {
+    name: 'id-match',
+    link: 'https://github.com/maddevsio/idmatch',
+    description: 'An open source tool enabling ID cards recognition to ensure security on your premises.',
+  },
+  {
+    name: 'your-cast',
+    link: 'https://github.com/maddevsio/yourcast.tv',
+    description: 'A solution to create lists of YouTube videos and stream them online in just a few seconds.',
+  },
+  {
+    name: 'heimdall',
+    link: 'https://github.com/maddevsio/heimdall',
+    description: 'A <span>security-monitoring</span> solution for Ethereum smart contracts.',
+  },
+]
