+++ client/pages/index.vue
@@ -15,7 +15,7 @@
 import AboutBanner from '@/components/core/banners/AboutBanner'
 import WeCreateItProjectsFor from '@/components/About/WeCreateItProjectsFor'
 import TechnologiesAndTools from '@/components/About/TechnologiesAndTools'
-import CaseStudies from '@/components/About/CaseStudies'
+import CaseStudies from '@/components/core/CaseStudies'
 import QuickProjectStart from '@/components/About/QuickProjectStart'
 import CustomerTestimonials from '@/components/About/CustomerTestimonials'
 import CustomerRates from '@/components/About/CustomerRates'
