+++ client/pages/projects.vue
@@ -9,7 +9,7 @@
 
 <script>
 import ProjectsBanner from '@/components/core/banners/ProjectsBanner'
-import CaseStudies from '@/components/About/CaseStudies'
+import CaseStudies from '@/components/core/CaseStudies'
 import Customers from '@/components/Projects/Customers'
 import OpenSource from '@/components/Projects/OpenSource'
 import { getMetadata, buildHead } from '@/data/seo'
